Fix compilation in gcc 9
Fixes https://github.com/isciences/exactextract/issues/82

+++ src/gdal_feature.h
@@ -134,7 +134,7 @@ class GDALFeature : public Feature
 
     void set(const std::string& name, std::size_t value) override
     {
-        if (value > std::numeric_limits<std::int64_t>::max()) {
+        if (value > static_cast<std::size_t>(std::numeric_limits<std::int64_t>::max())) {
             throw std::runtime_error("Value too large to write");
         }
         OGR_F_SetFieldInteger64(m_feature, field_index(name), static_cast<std::int64_t>(value));
